--- src/main/java/org/example/oshipserver/domain/admin/dto/request/RequestRateDto.java
@@ -1,5 +1,0 @@
-package org.example.oshipserver.domain.admin.dto.request;
-
-public class RequestRateDto {
-
-}
