+++ src/main/java/org/example/oshipserver/domain/admin/service/AdminService.java
@@ -3,7 +3,7 @@ package org.example.oshipserver.domain.admin.service;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.oshipserver.domain.admin.dto.RequestZone;
+import org.example.oshipserver.domain.admin.dto.request.RequestZone;
 import org.example.oshipserver.domain.admin.dto.request.RateCreateRequest;
 import org.example.oshipserver.domain.admin.dto.response.ResponseRateDto;
 import org.example.oshipserver.domain.carrier.service.AdminCarrierService;
